pyproject.toml
[project]
name = "waseller-crew"
version = "0.1.0"
description = "Microservicio CrewAI para shadow-compare con Waseller"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.10.0",
    "crewai>=0.86.0",
    "python-dotenv>=1.0.0",
    "ftfy>=6.2.0",
]

[project.optional-dependencies]
dev = [
    "httpx>=0.28.0",
    "pytest>=8.0.0",
    "ruff>=0.8.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/crew_shadow_crewai"]

[tool.ruff]
line-length = 100
target-version = "py312"
